General refactor/restructure of the project
