+++ .github/workflows/lock.yaml
@@ -4,7 +4,7 @@ name: Lock
 # yamllint disable-line rule:truthy
 on:
   schedule:
-    - cron: "0 5 * * *"
+    - cron: "0 17 * * *"
   workflow_dispatch:
 
 jobs:
