--- .github/workflows/requirements.txt
@@ -1,2 +1,0 @@
-pip==23.0.1
-poetry==1.4.1
