+++ .github/workflows/stale.yaml
@@ -4,7 +4,7 @@ name: Stale
 # yamllint disable-line rule:truthy
 on:
   schedule:
-    - cron: "0 4 * * *"
+    - cron: "0 16 * * *"
   workflow_dispatch:
 
 jobs:
