+++ .gitignore
@@ -116,8 +116,12 @@ venv.bak/
 
 # mypy
 .mypy_cache/
-.dmypy.json
-dmypy.json
+
+# Visual Studio Code
+.vscode
+
+# ruff
+.ruff_cache
 
 # Pyre type checker
 .pyre/
