+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Examples for this library."""
