+++ examples/ruff.toml
@@ -1,0 +1,6 @@
+# This extend our general Ruff rules specifically for the examples
+extend = "../pyproject.toml"
+
+extend-ignore = [
+  "T201", # Allow the use of print() in examples
+]
