+++ omnikinverter/models.py
@@ -41,18 +41,21 @@ class Inverter:
         """Return Inverter object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The JSON data from the Omnik Inverter.
 
         Returns:
+        -------
             An Inverter object.
 
         Raises:
+        ------
             OmnikInverterWrongValuesError: Inverter pass on
                 incorrect data (day and total are equal).
         """
 
         def get_value(search: str) -> Any:
-            if data[search] != "":
+            if data[search]:
                 return data[search]
             return None
 
@@ -60,18 +63,18 @@ class Inverter:
             """Check if the values are not equal to each other.
 
             Args:
+            ----
                 data_list: List of values to check.
 
             Returns:
+            -------
                 Boolean value.
             """
-            res = all(ele == data_list[0] for ele in data_list)
-            return res
+            return all(ele == data_list[0] for ele in data_list)
 
         if validation([data["i_eday"], data["i_eall"]]):
-            raise OmnikInverterWrongValuesError(
-                "Inverter pass on incorrect data (day and total are equal)"
-            )
+            msg = "Inverter pass on incorrect data (day and total are equal)"
+            raise OmnikInverterWrongValuesError(msg)
 
         return Inverter(
             serial_number=get_value("i_sn"),
@@ -90,9 +93,11 @@ class Inverter:
         """Return Inverter object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The HTML (webscraping) data from the Omnik Inverter.
 
         Returns:
+        -------
             An Inverter object.
         """
 
@@ -102,7 +107,7 @@ class Inverter:
                     re.Match[str],
                     re.search(f'(?<={search_key}=").*?(?=";)', data.replace(" ", "")),
                 ).group(0)
-                if match != "":
+                if match:
                     if search_key in ["webdata_now_p", "webdata_rate_p"]:
                         return int(match)
                     if search_key in ["webdata_today_e", "webdata_total_e"]:
@@ -110,9 +115,8 @@ class Inverter:
                     return match
                 return None
             except AttributeError as exception:
-                raise OmnikInverterWrongSourceError(
-                    "Your inverter has no data source from a html file."
-                ) from exception
+                msg = "Your inverter has no data source from a html file."
+                raise OmnikInverterWrongSourceError(msg) from exception
 
         return Inverter(
             serial_number=get_value("webdata_sn"),
@@ -131,9 +135,11 @@ class Inverter:
         """Return Inverter object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The JS (webscraping) data from the Omnik Inverter.
 
         Returns:
+        -------
             An Inverter object.
         """
 
@@ -155,7 +161,7 @@ class Inverter:
                         .split(",")
                     )
 
-                if matches[position] != "":
+                if matches[position]:
                     if position in [4, 5, 6, 7]:
                         if position in [4, 5]:
                             return int(matches[position].replace(" ", ""))
@@ -168,9 +174,8 @@ class Inverter:
                     return matches[position].replace(" ", "")
                 return None
             except AttributeError as exception:
-                raise OmnikInverterWrongSourceError(
-                    "Your inverter has no data source from a javascript file."
-                ) from exception
+                msg = "Your inverter has no data source from a javascript file."
+                raise OmnikInverterWrongSourceError(msg) from exception
 
         return Inverter(
             serial_number=get_value(0),
@@ -189,12 +194,13 @@ class Inverter:
         """Return Inverter object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The binary data from the Omnik Inverter.
 
         Returns:
+        -------
             An Inverter object.
         """
-
         return Inverter(
             **data,
             model=None,  # Not able to deduce this from raw message yet
@@ -218,9 +224,11 @@ class Device:
         """Return Device object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The JSON data from the Omnik Inverter.
 
         Returns:
+        -------
             An Device object.
         """
         return Device(
@@ -234,21 +242,23 @@ class Device:
         """Return Device object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The HTML (webscraping) data from the Omnik Inverter.
 
         Returns:
+        -------
             An Device object.
         """
-
         for correction in [" ", "%"]:
             data = data.replace(correction, "")
 
         def get_value(search_key: str) -> Any:
             match = cast(
-                re.Match[str], re.search(f'(?<={search_key}=").*?(?=";)', data)
+                re.Match[str],
+                re.search(f'(?<={search_key}=").*?(?=";)', data),
             ).group(0)
 
-            if match != "":
+            if match:
                 if search_key in ["cover_sta_rssi"]:
                     return int(match)
                 return match
@@ -265,9 +275,11 @@ class Device:
         """Return Device object from the Omnik Inverter response.
 
         Args:
+        ----
             data: The JS (webscraping) data from the Omnik Inverter.
 
         Returns:
+        -------
             An Device object.
         """
         for correction in [" ", "%"]:
@@ -275,10 +287,11 @@ class Device:
 
         def get_value(search_key: str) -> Any:
             match = cast(
-                re.Match[str], re.search(f'(?<={search_key}=").*?(?=";)', data)
+                re.Match[str],
+                re.search(f'(?<={search_key}=").*?(?=";)', data),
             ).group(0)
 
-            if match != "":
+            if match:
                 if search_key == "m2mRssi":
                     return int(match)
                 return match
