+++ omnikinverter/omnikinverter.py
@@ -3,13 +3,11 @@ from __future__ import annotations
 
 import asyncio
 import json
-from collections.abc import Mapping
 from dataclasses import dataclass
 from typing import Any
 
-import aiohttp
 import async_timeout
-from aiohttp.client import ClientError, ClientResponseError, ClientSession
+from aiohttp import BasicAuth, ClientError, ClientResponseError, ClientSession
 from aiohttp.hdrs import METH_GET
 from yarl import URL
 
@@ -44,20 +42,23 @@ class OmnikInverter:
         uri: str,
         *,
         method: str = METH_GET,
-        params: Mapping[str, Any] | None = None,
+        params: dict[str, Any] | None = None,
     ) -> str:
         """Handle a request to a Omnik Inverter device.
 
         Args:
+        ----
             uri: Request URI, without '/', for example, 'status'
             method: HTTP Method to use.
             params: Extra options to improve or limit the response.
 
         Returns:
+        -------
             A Python dictionary (text) with the response from
             the Omnik Inverter.
 
         Raises:
+        ------
             OmnikInverterConnectionError: An error occurred while communicating
                 with the Omnik Inverter.
             OmnikInverterAuthError: Authentication failed with the Omnik Inverter.
@@ -78,12 +79,11 @@ class OmnikInverter:
             if self.source_type == "html" and (
                 self.username is None or self.password is None
             ):
-                raise OmnikInverterAuthError(
-                    "A username and/or password is missing from the request"
-                )
+                msg = "A username and/or password is missing from the request"
+                raise OmnikInverterAuthError(msg)
             auth = None
             if self.username and self.password:
-                auth = aiohttp.BasicAuth(self.username, self.password)
+                auth = BasicAuth(self.username, self.password)
 
             try:
                 async with async_timeout.timeout(self.request_timeout):
@@ -96,20 +96,19 @@ class OmnikInverter:
                     )
                     response.raise_for_status()
             except asyncio.TimeoutError as exception:
-                raise OmnikInverterConnectionError(
-                    "Timeout occurred while connecting to Omnik Inverter device"
-                ) from exception
+                msg = "Timeout occurred while connecting to Omnik Inverter device"
+                raise OmnikInverterConnectionError(msg) from exception
             except (ClientError, ClientResponseError) as exception:
-                raise OmnikInverterConnectionError(
-                    "Error occurred while communicating with Omnik Inverter device"
-                ) from exception
+                msg = "Error occurred while communicating with Omnik Inverter device"
+                raise OmnikInverterConnectionError(msg) from exception
 
             types = ["application/json", "application/x-javascript", "text/html"]
             content_type = response.headers.get("Content-Type", "")
             if not any(item in content_type for item in types):
                 text = await response.text()
+                msg = "Unexpected response from the Omnik Inverter device"
                 raise OmnikInverterError(
-                    "Unexpected response from the Omnik Inverter device",
+                    msg,
                     {"Content-Type": content_type, "response": text},
                 )
 
@@ -125,18 +124,21 @@ class OmnikInverter:
     async def tcp_request(self) -> dict[str, Any]:
         """Perform a raw TCP request to the Omnik device.
 
-        Returns:
+        Returns
+        -------
             A Python dictionary (text) with the response from
             the Omnik Inverter.
 
-        Raises:
+        Raises
+        ------
             OmnikInverterAuthError: Serial number is required to communicate
                 with the Omnik Inverter.
             OmnikInverterConnectionError: An error occurred while communicating
                 with the Omnik Inverter.
         """
         if self.serial_number is None:
-            raise OmnikInverterAuthError("serial_number is missing from the request")
+            msg = "serial_number is missing from the request"
+            raise OmnikInverterAuthError(msg)
 
         try:
             if self._socket_mock is not None:
@@ -144,9 +146,8 @@ class OmnikInverter:
             else:  # pragma: no cover
                 reader, writer = await asyncio.open_connection(self.host, self.tcp_port)
         except OSError as exception:
-            raise OmnikInverterConnectionError(
-                "Failed to open a TCP connection to the Omnik Inverter device"
-            ) from exception
+            msg = "Failed to open a TCP connection to the Omnik Inverter device"
+            raise OmnikInverterConnectionError(msg) from exception
 
         try:
             writer.write(tcp.create_information_request(self.serial_number))
@@ -158,19 +159,20 @@ class OmnikInverter:
             try:
                 await writer.wait_closed()
             except OSError as exception:
-                raise OmnikInverterConnectionError(
-                    "Failed to communicate with the Omnik Inverter device over TCP"
-                ) from exception
+                msg = "Failed to close the TCP connection to the Omnik Inverter device"
+                raise OmnikInverterConnectionError(msg) from exception
 
         return tcp.parse_messages(self.serial_number, raw_msg)
 
     async def inverter(self) -> Inverter:
         """Get values from your Omnik Inverter.
 
-        Returns:
+        Returns
+        -------
             A Inverter data object from the Omnik Inverter.
 
-        Raises:
+        Raises
+        ------
             OmnikInverterError: Unknown source type.
         """
         if self.source_type == "json":
@@ -186,15 +188,18 @@ class OmnikInverter:
             fields = await self.tcp_request()
             return Inverter.from_tcp(fields)
 
-        raise OmnikInverterError(f"Unknown source type `{self.source_type}`")
+        msg = f"Unknown source type `{self.source_type}`"
+        raise OmnikInverterError(msg)
 
     async def device(self) -> Device:
         """Get values from the device.
 
-        Returns:
+        Returns
+        -------
             A Device data object from the Omnik Inverter. None on the "tcp" source_type.
 
-        Raises:
+        Raises
+        ------
             OmnikInverterError: Unknown source type.
         """
         if self.source_type == "json":
@@ -210,7 +215,8 @@ class OmnikInverter:
             # None of the fields are available through a TCP data dump.
             return Device()
 
-        raise OmnikInverterError(f"Unknown source type `{self.source_type}`")
+        msg = f"Unknown source type `{self.source_type}`"
+        raise OmnikInverterError(msg)
 
     async def close(self) -> None:
         """Close open client session."""
@@ -220,7 +226,8 @@ class OmnikInverter:
     async def __aenter__(self) -> OmnikInverter:
         """Async enter.
 
-        Returns:
+        Returns
+        -------
             The Omnik Inverter object.
         """
         return self
@@ -229,6 +236,7 @@ class OmnikInverter:
         """Async exit.
 
         Args:
+        ----
             _exc_info: Exec type.
         """
         await self.close()
