+++ pyproject.toml
@@ -26,45 +26,31 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9"
 aiohttp = ">=3.0.0"
+python = "^3.9"
 yarl = ">=1.6.0"
-asynctest = "^0.13.0"
-
-[tool.poetry.group.dev.dependencies]
-aresponses = "^2.1.6"
-bandit = "^1.7.4"
-black = ">=22.10,<24.0"
-blacken-docs = "^1.12.1"
-codespell = "^2.2.2"
-coverage = {version = "^6.5", extras = ["toml"]}
-darglint = "^1.8.1"
-flake8 = "^4.0.1"
-flake8-bandit = "^3.0.0"
-flake8-bugbear = "^22.10.27"
-flake8-builtins = "^2.0.1"
-flake8-comprehensions = "^3.10.1"
-flake8-docstrings = "^1.5.0"
-flake8-eradicate = "^1.4.0"
-flake8-markdown = "^0.3.0"
-flake8-simplify = "^0.19.3"
-isort = "^5.10.1"
-mypy = "^0.991"
-pre-commit = ">=2.20,<4.0"
-pre-commit-hooks = "^4.3.0"
-pylint = "^2.15.6"
-pytest = "^7.2.0"
-pytest-asyncio = "^0.20.2"
-pytest-cov = "^4.0.0"
-pyupgrade = "^3.2.2"
-safety = "^2.3.2"
-vulture = "^2.6"
-yamllint = "^1.28.0"
-types-cachetools = "^5.2.1"
 
 [tool.poetry.urls]
 "Bug Tracker" = "https://github.com/klaasnicolaas/python-omnikinverter/issues"
 Changelog = "https://github.com/klaasnicolaas/python-omnikinverter/releases"
+
+[tool.poetry.group.dev.dependencies]
+ruff = ">=0.0.243,<0.0.260"
+aresponses = "^2.1.6"
+black = ">=22.10,<24.0"
+blacken-docs = "^1.13.0"
+codespell = "^2.2.4"
+coverage = {version = ">=7.2,<8.0", extras = ["toml"]}
+mypy = ">=1.0,<1.2"
+pre-commit = ">=3.2,<4.0"
+pre-commit-hooks = "^4.4.0"
+pylint = "^2.17.1"
+pytest = "^7.2.2"
+pytest-asyncio = "^0.21.0"
+pytest-cov = "^4.0.0"
+yamllint = "^1.30.0"
+asynctest = "^0.13.0"
+covdefaults = "^2.3.0"
 
 [tool.black]
 target-version = ['py39']
@@ -72,27 +58,16 @@ target-version = ['py39']
 [tool.coverage.paths]
 source = ["omnikinverter"]
 
-[tool.coverage.report]
-show_missing = true
-exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]
-
 [tool.coverage.run]
-branch = true
+plugins = ["covdefaults"]
 source = ["omnikinverter"]
-
-[tool.isort]
-profile = "black"
-multi_line_output = 3
 
 [tool.mypy]
 # Specify the target platform details in config, so your developers are
 # free to run mypy on Windows, Linux, or macOS and get consistent
 # results.
 platform = "linux"
-python_version = 3.9
-
-# flake8-mypy expects the two following for sensible formatting
-show_column_numbers = true
+python_version = "3.9"
 
 # show error messages from unrelated files
 follow_imports = "normal"
@@ -107,7 +82,7 @@ disallow_incomplete_defs = true
 disallow_subclassing_any = true
 disallow_untyped_calls = true
 disallow_untyped_defs = true
-disallow_untyped_decorators = false # thanks backoff :(
+disallow_untyped_decorators = true
 no_implicit_optional = true
 no_implicit_reexport = true
 strict_optional = true
@@ -118,8 +93,13 @@ warn_return_any = true
 warn_unused_configs = true
 warn_unused_ignores = true
 
-# No incremental mode
-cache_dir = "/dev/null"
+[tool.pylint.MASTER]
+extension-pkg-whitelist = [
+  "pydantic"
+]
+ignore= [
+  "tests"
+]
 
 [tool.pylint.BASIC]
 good-names = [
@@ -135,6 +115,9 @@ good-names = [
     "T",
 ]
 
+[tool.pylint.DESIGN]
+max-attributes=20
+
 [tool.pylint."MESSAGES CONTROL"]
 disable= [
     "too-few-public-methods",
@@ -149,17 +132,30 @@ ignore-imports = true
 [tool.pylint.FORMAT]
 max-line-length=88
 
-[tool.pylint.DESIGN]
-max-attributes=20
-
 [tool.pytest.ini_options]
 addopts = "--cov"
 asyncio_mode = "auto"
 
-[tool.vulture]
-min_confidence = 80
-paths = ["omnikinverter"]
-verbose = true
+[tool.ruff]
+select = ["ALL"]
+ignore = [
+  "ANN101", # Self... explanatory
+  "ANN401", # Opinioated warning on disallowing dynamically typed expressions
+  "D203", # Conflicts with other rules
+  "D213", # Conflicts with other rules
+  "D417", # False positives in some occasions
+  "PLR2004", # Just annoying, not really useful
+]
+
+[tool.ruff.flake8-pytest-style]
+mark-parentheses = false
+fixture-parentheses = false
+
+[tool.ruff.isort]
+known-first-party = ["omnikinverter"]
+
+[tool.ruff.mccabe]
+max-complexity = 25
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
