+++ tests/__init__.py
@@ -1,18 +1,16 @@
 """Testing functionality for the Omnik Inverter device."""
-import os
+from pathlib import Path
 
 
-def _fixture_path(filename: str) -> str:
-    return os.path.join(os.path.dirname(__file__), "fixtures", filename)
+def _fixture_path(filename: str) -> Path:
+    return Path(__file__).parent / "fixtures" / filename
 
 
 def load_fixtures(filename: str) -> str:
     """Load a fixture."""
-    with open(_fixture_path(filename), encoding="utf-8") as fptr:
-        return fptr.read()
+    return _fixture_path(filename).read_text(encoding="utf-8")
 
 
 def load_fixture_bytes(filename: str) -> bytes:
     """Load a fixture as binary data."""
-    with open(_fixture_path(filename), "rb") as fptr:
-        return fptr.read()
+    return _fixture_path(filename).read_bytes()
